#44 [refactor] 게시글 aggregate dto. vo 코드 리팩토링
수정으로 인해 해당 파일들의 이름이 변경과 삭제가 있었습니다.
