--- OnionHotSayYo/src/main/java/org/omoknoone/onionhotsayyo/post/command/vo/ResponsePostListByCategory.java
@@ -1,4 +1,0 @@
-package org.omoknoone.onionhotsayyo.post.command.vo;
-
-public class ResponsePostListByCategory {
-}
